libjlm/ir: rename getelementpts.hpp/*.cpp to GetElementPtr.hpp/*.cpp
