+++ libjlm/include/jlm/ir/operators.hpp
@@ -10,7 +10,7 @@
 #include <jlm/ir/operators/call.hpp>
 #include <jlm/ir/operators/delta.hpp>
 #include <jlm/ir/operators/gamma.hpp>
-#include <jlm/ir/operators/getelementptr.hpp>
+#include <jlm/ir/operators/GetElementPtr.hpp>
 #include <jlm/ir/operators/lambda.hpp>
 #include <jlm/ir/operators/load.hpp>
 #include <jlm/ir/operators/operators.hpp>
